pyproject.toml
[project]
name = "mcp-gateway"
version = "0.1.0"
description = "MCP Gateway - Aggregate multiple MCP servers into one endpoint"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.17.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.0.0",
    "anyio>=4.0.0",
    "httpx>=0.27.0",
    "starlette>=0.45.0",
    "prometheus-client>=0.21.0",
    "watchdog>=4.0.0",
    "jinja2>=3.1.0",
    "aiofiles>=24.0.0",
    "structlog>=24.0.0",
    "pybreaker>=1.2.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "locust>=2.24.0",
    "respx>=0.21.0",
    "aioresponses>=0.7.0",
]

[project.scripts]
mcp-gateway = "mcp_gateway.main:main"
mcp-gateway-approve = "mcp_gateway.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "prometheus_client.*",
    "watchdog.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "benchmark: marks tests as benchmark tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
